Compensation에 File업데이트 및 File 조회

+++ src/main/java/com/coway/trust/biz/common/impl/FileServiceImpl.java
@@ -124,8 +124,10 @@ public class FileServiceImpl implements FileService {
 		FileVO fileVO = this.getFile(fileId);
 
 		try {
+			if(fileVO != null){
 			FileUtils.forceDelete(new File(
 					baseDir + File.separator + fileVO.getFileSubPath() + File.separator + fileVO.getPhysiclFileName()));
+			}
 		} catch (IOException e) {
 			LOGGER.error("deleteFile Fail : {}", e.getMessage());
 			// throw new ApplicationException(e);
